% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMeta.R
\name{compareMeta}
\alias{compareMeta}
\title{Compare metadata between two REDCap projects}
\usage{
compareMeta(key1, key2)
}
\arguments{
\item{key1}{the API_KEY of the REDCap project used as reference.}

\item{key2}{the API_KEY of the REDCap project to examine for differences versus the reference.}
}
\value{
a string containing a Markdown report of differences
}
\description{
Given two REDCap project API_KEYs return a Markdown report of all discovered
differences.
}
