% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMeta.R
\name{exportMeta}
\alias{exportMeta}
\title{Retrieve metadata for a REDCap project}
\usage{
exportMeta(key)
}
\arguments{
\item{key}{the API_KEY of the REDCap project}
}
\value{
a structure representing all the meta data
}
\description{
Given a key to a REDCap project return a structure containing all the
meta data (e.g. Data Dictionary) of that project.
}
