% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMeta.R
\name{importMeta}
\alias{importMeta}
\title{Update a REDCap project to reflect a provided meta definition.}
\usage{
importMeta(key, meta)
}
\arguments{
\item{key}{the API_KEY of the REDCap project to update.}

\item{meta}{the meta definition}
}
\value{
logical; The success or failure of the update.
}
\description{
Update a REDCap project to be the current meta definition. Useful for keeping
a development and production REDCap project synchronized.
}
\seealso{
exportMeta
}
